+++ packages/pwa-kit-dev/CHANGELOG.md
@@ -3,6 +3,8 @@
 
 - Bugfix: resolve performance issue due to webpack stats [#1418](https://github.com/SalesforceCommerceCloud/pwa-kit/pull/1418)
 
+- Added a new `--wait` flag to the `push` command. When `--target` is specified, setting the `--wait` flag will cause the script to exit after the triggered deployment has completed, rather than after the bundle has been pushed.
+
 ## v3.1.0 (Jul 26, 2023)
 
 - Allow setting cookies using the `MRT_ALLOW_COOKIES` environment variable [#1318](https://github.com/SalesforceCommerceCloud/pwa-kit/pull/1318)